# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ticket Management System API",
        "description": "API description",
        "contact": {
            "email": "utkarsh.wilson@regur.net"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get all projects",
                "description": "Get all projects",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "All projects fetched successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Error in fetching projects"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/projects/new": {
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Create a new project",
                "description": "Create a new project",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "title": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "Project title",
                                    "description": "Project description"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Project created successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Error in creating project"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/projects/{projectId}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get a projects",
                "description": "Get a projects",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "ID of Project",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All projects details fetched successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Error in fetching user details"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/projects/edit/{projectId}": {
            "put": {
                "tags": [
                    "Projects"
                ],
                "summary": "Update a project",
                "description": "Update a project",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "ID of Project",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "title": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "New Project title",
                                    "description": "New Project description"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Project updated successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Error in updating project"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/projects/delete/{projectId}": {
            "delete": {
                "tags": [
                    "Projects"
                ],
                "summary": "Delete a project",
                "description": "Delete a project",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "ID of Project",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project deleted successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Error in deleting project"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Login an existing user",
                "description": "Login an existing user",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "email": {
                                                "type": "string"
                                            },
                                            "password": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "example@email.com",
                                    "password": "abc123"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "400": {
                        "description": "Error in loggin in"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "Register a new user",
                "description": "Register a new user with email and password",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "password": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "username",
                                    "email": "example@email.com",
                                    "password": "abc123"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User registered successfully"
                    },
                    "400": {
                        "description": "Error in registering user"
                    }
                }
            }
        },
        "/api/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get user details",
                "description": "Get user details",
                "operationId": "userDetails",
                "responses": {
                    "200": {
                        "description": "User details fetched successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Error in fetching user details"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "name": "Authorization"
            }
        }
    }
}